Replace direct dot-product DCT/IDCT with verified exact fast-DCT kernel
Adds FastScalarInverseDct/FastScalarForwardDct, an even/odd butterfly
factorization derived from the ITU-T.81 A.3.3 definition (21 multiplies
per 1D pass vs. the direct definition's 64), and makes it the
DctKernelSelector default. Numerically verified via matrix cross-check
and impulse-response tests independent of the scalar reference kernel,
and benchmarked 19-36% faster per-block than the previous Vector256
production default with zero regressions across 2,260 corpus files.

IInverseDctKernel/IForwardDctKernel gain PrepareDequantTable/
PrepareQuantTable default-interface methods so a kernel can fold
per-frequency scale factors into the (de)quant table once per
component instead of once per block; existing kernels use the default
identity implementation unchanged.

True minimal-multiply AAN (the classical ~5-multiply-per-pass odd
branch) could not be safely completed this session without risking a
silent scale-factor bug, despite independently verifying its scale
factors against three sources (AanScaleFactors.cs, ready but unused).
Follow-up: https://github.com/jhaygood86/PeachImage/issues/5

+++ src/PeachImage/Formats/Jpeg/Dct/ScalarInverseDct.cs
@@ -10,7 +10,7 @@ internal sealed class ScalarInverseDct : IInverseDctKernel
     private static readonly double[,] Cosines = BuildCosineTable();
     private static readonly double[] Normalization = [1.0 / 1.4142135623730951, 1, 1, 1, 1, 1, 1, 1];
 
-    public void Transform(ReadOnlySpan<short> coefficients, ReadOnlySpan<ushort> dequantTable, Span<byte> output, int outputStride)
+    public void Transform(ReadOnlySpan<short> coefficients, ReadOnlySpan<float> dequantTable, Span<byte> output, int outputStride)
     {
         Span<double> dequantized = stackalloc double[64];
         for (int i = 0; i < 64; i++)
