+++ src/PeachImage/Formats/Jpeg/Dct/Vector128InverseDct.cs
@@ -11,12 +11,12 @@ internal sealed class Vector128InverseDct : IInverseDctKernel
 {
     private static readonly (Vector128<float> Lo, Vector128<float> Hi)[] WeightedCosineRows = BuildRows();
 
-    public void Transform(ReadOnlySpan<short> coefficients, ReadOnlySpan<ushort> dequantTable, Span<byte> output, int outputStride)
+    public void Transform(ReadOnlySpan<short> coefficients, ReadOnlySpan<float> dequantTable, Span<byte> output, int outputStride)
     {
         Span<float> dequantized = stackalloc float[64];
         for (int i = 0; i < 64; i++)
         {
-            dequantized[i] = coefficients[i] * (float)dequantTable[i];
+            dequantized[i] = coefficients[i] * dequantTable[i];
         }
 
         Span<float> rowPass = stackalloc float[64];
